refactor(backend): move JoinSessionCallbacks/JoinSessionOptions to types.ts (review)
Centralise all RoomManager plumbing types in types.ts alongside
RoomManagerDeps, per review feedback on #3360 — JoinSessionCallbacks and
JoinSessionOptions were previously defined in client-lifecycle.ts, which
undermined the point of the refactor. types.ts takes a type-only import of
SessionLeaveResult from client-lifecycle.ts (erased at compile time, so no
runtime circular dependency).
